+++ amundsen_application/static/js/components/common/TagsList/tests/index.spec.tsx
@@ -114,7 +114,7 @@ describe('mapStateToProps', () => {
   let expectedOtherTags;
   beforeEach(() => {
     result = mapStateToProps(globalState);
-    const allTags = globalState.allTags.allTags;
+    const allTags = globalState.tags.allTags;
     const curatedTagsList = getCuratedTags();
     expectedCuratedTags = allTags.filter((tag) => curatedTagsList.indexOf(tag.tag_name) !== -1);
     expectedOtherTags = allTags.filter((tag) => curatedTagsList.indexOf(tag.tag_name) === -1);
@@ -129,6 +129,6 @@ describe('mapStateToProps', () => {
   });
 
   it('sets isLoading on the props', () => {
-    expect(result.isLoading).toEqual(globalState.allTags.isLoading);
+    expect(result.isLoading).toEqual(globalState.tags.isLoadingAllTags);
   });
 });
